PAY-553 Retreive payment status from govPay

+++ api/src/test/java/uk/gov/hmcts/payment/api/componenttests/CardPaymentComponentTest.java
@@ -31,7 +31,7 @@ public class CardPaymentComponentTest extends ComponentTestBase {
         assertNotNull(paymentFeeLink);
         assertEquals(paymentFeeLink.getPayments().size(), 1);
         assertEquals(paymentFeeLink.getFees().size(), 1);
-        assertEquals(paymentFeeLink.getPayments().get(0).getAmount(),  new Integer(100));
+        assertEquals(paymentFeeLink.getPayments().get(0).getAmount(),  new Integer(1000000));
         assertEquals(paymentFeeLink.getFees().get(0).getCode(), "X0033");
     }
